+++ middleware/signer_required.go
@@ -11,9 +11,9 @@ import (
 func SignerRequired() gin.HandlerFunc {
 	return func(c *gin.Context) {
 		conf := c.MustGet("config").(*config.Config)
-		ghID := c.MustGet("gh_id").(int)
+		ghID := c.MustGet("gh_id").(int64)
 
-		if int64(ghID) != conf.SignerGitHubID {
+		if ghID != conf.SignerGitHubID {
 			c.AbortWithStatus(http.StatusForbidden)
 			return
 		}
